+++ src/metador_core/cli/__init__.py
@@ -1,0 +1,7 @@
+"""Metador CLI for system introspection."""
+import typer
+
+from . import general
+
+app = typer.Typer()
+app.add_typer(general.app, name="self")