+++ CHANGELOG.md
@@ -6,7 +6,12 @@ Please consult the changelog to inform yourself about breaking changes and secur
 
 ## [v0.1.1](https://github.com/Materials-Data-Science-and-Informatics/metador-core/tree/v0.1.1) <small>(2023-??-??)</small> { id="0.1.1" }
 
+* added `metador` CLI tools for self-diagnosis (other features can be added in the future)
 * added CSV/TSV widget
+* added example schemas for tutorial in material science context
+* renamed `embed_file` to `pack_file` and simplified its arguments
+* removed `creator` from `BibMeta` (author and creator are equivalent in schema.org)
+* minor tweaks and fixes
 
 ## [v0.1.0](https://github.com/Materials-Data-Science-and-Informatics/metador-core/tree/v0.1.0) <small>(2023-08-08)</small> { id="0.1.0" }